Move model manager and mode cache to their own module
